pyproject.toml
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "fastapi-microservice-py"
version = "0.1.0"
description = "A POC for a maintainable FastAPI microservice in Python."
authors = ["u-ways <work@u-ways.info>"]
repository = "https://github.com/u-ways/fastapi-microservice-py"
readme = "README.md"
license = "MIT"
packages = [ { include = "src", from = "." } ]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.110.1"
uvicorn = "^0.29.0"
python-dotenv = "^1.0.1"
prometheus-fastapi-instrumentator = "^7.0.0"

[tool.poetry.dev-dependencies]
pyclean = "^3.0.0"
black = "^24.4.0"
flake8 = "^7.0.0"
flake8-pyproject = "^1.2.3"
httpx = "^0.27.0"
pytest = "^8.1.1"
parameterized = "^0.9.0"

[tool.black]
line-length = 120

[tool.flake8]
max-line-length = 120
count = true

[tool.pytest.ini_options]
pythonpath = "./src"
testpaths = "tests"